+++ app/src/main/java/com/comppot/mindsnack/ui/components/Messages.kt
@@ -1,0 +1,27 @@
+package com.comppot.mindsnack.ui.components
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+
+@Composable
+fun EmptyListMessage(message: String) {
+    Box(modifier = Modifier.fillMaxSize()) {
+        Text(
+            message,
+            color = MaterialTheme.colorScheme.onSurfaceVariant,
+            modifier = Modifier.align(Alignment.Center)
+        )
+    }
+}
+
+@Preview
+@Composable
+private fun EmptyListMessagePreview() {
+    EmptyListMessage("No results")
+}
