+++ app/src/main/java/com/comppot/mindsnack/ui/screens/article/Rating.kt
@@ -1,6 +1,6 @@
 @file:OptIn(ExperimentalMaterial3Api::class)
 
-package com.comppot.mindsnack.ui.components
+package com.comppot.mindsnack.ui.screens.article
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.interaction.MutableInteractionSource
