+++ app/src/main/java/com/comppot/mindsnack/ui/screens/tab/home/HomeScreen.kt
@@ -1,19 +1,14 @@
 package com.comppot.mindsnack.ui.screens.tab.home
 
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.LazyRow
 import androidx.compose.foundation.lazy.items
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
@@ -22,8 +17,9 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import com.comppot.mindsnack.R
 import com.comppot.mindsnack.model.Article
 import com.comppot.mindsnack.model.Category
-import com.comppot.mindsnack.ui.components.ArticleCard
+import com.comppot.mindsnack.ui.components.ArticleItem
 import com.comppot.mindsnack.ui.components.CategoryItem
+import com.comppot.mindsnack.ui.components.EmptyListMessage
 import com.comppot.mindsnack.ui.components.FullScreenLoading
 
 @Composable
@@ -34,7 +30,7 @@ fun HomeScreen(openArticle: (Long) -> Unit = {}, viewModel: HomeViewModel = hilt
         CategoryList(state.categories, state.selectedCategory, viewModel::selectCategory)
         when {
             state.isLoading -> FullScreenLoading()
-            state.articles.isEmpty() -> NoArticles()
+            state.articles.isEmpty() -> EmptyListMessage(stringResource(R.string.home_screen_no_articles))
             else -> ArticleList(state.articles, openArticle)
         }
     }
@@ -63,19 +59,8 @@ private fun ArticleList(articles: List<Article>, openArticle: (Long) -> Unit) {
         verticalArrangement = Arrangement.spacedBy(16.dp)
     ) {
         items(articles) {
-            ArticleCard(it, Modifier.fillMaxWidth(), openArticle)
+            ArticleItem(it, Modifier.fillMaxWidth(), openArticle)
         }
-    }
-}
-
-@Composable
-private fun NoArticles() {
-    Box(modifier = Modifier.fillMaxSize()) {
-        Text(
-            stringResource(id = R.string.home_screen_no_articles),
-            color = MaterialTheme.colorScheme.onSurfaceVariant,
-            modifier = Modifier.align(Alignment.Center)
-        )
     }
 }
