+++ app/src/main/java/com/comppot/mindsnack/ui/screens/tab/saved/SavedScreen.kt
@@ -1,24 +1,14 @@
 package com.comppot.mindsnack.ui.screens.tab.saved
 
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.lazy.grid.GridCells
-import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
-import androidx.compose.foundation.lazy.grid.items
 import androidx.compose.foundation.lazy.staggeredgrid.LazyVerticalStaggeredGrid
 import androidx.compose.foundation.lazy.staggeredgrid.StaggeredGridCells
 import androidx.compose.foundation.lazy.staggeredgrid.items
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
@@ -26,8 +16,9 @@ import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.comppot.mindsnack.R
 import com.comppot.mindsnack.model.Article
+import com.comppot.mindsnack.ui.components.EmptyListMessage
 import com.comppot.mindsnack.ui.components.FullScreenLoading
-import com.comppot.mindsnack.ui.components.SavedArticleCard
+import com.comppot.mindsnack.ui.components.SavedArticleItem
 
 @Composable
 fun SavedScreen(openArticle: (Long) -> Unit = {}, viewModel: SavedViewModel = hiltViewModel()) {
@@ -35,7 +26,7 @@ fun SavedScreen(openArticle: (Long) -> Unit = {}, viewModel: SavedViewModel = hi
 
     when {
         articles == null -> FullScreenLoading()
-        articles.isEmpty() -> NoArticles()
+        articles.isEmpty() -> EmptyListMessage(stringResource(R.string.saved_screen_no_articles))
         else -> ArticleList(articles, openArticle)
     }
 }
@@ -50,19 +41,8 @@ private fun ArticleList(articles: List<Article>, openArticle: (Long) -> Unit) {
         modifier = Modifier.fillMaxSize()
     ) {
         items(articles) {
-            SavedArticleCard(it, Modifier.fillMaxWidth(), openArticle)
+            SavedArticleItem(it, Modifier.fillMaxWidth(), openArticle)
         }
-    }
-}
-
-@Composable
-private fun NoArticles() {
-    Box(modifier = Modifier.fillMaxSize()) {
-        Text(
-            stringResource(id = R.string.saved_screen_no_articles),
-            color = MaterialTheme.colorScheme.onSurfaceVariant,
-            modifier = Modifier.align(Alignment.Center)
-        )
     }
 }
