+++ app/src/main/java/com/comppot/mindsnack/ui/screens/tab/search/SearchScreen.kt
@@ -2,11 +2,9 @@ package com.comppot.mindsnack.ui.screens.tab.search
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
@@ -34,8 +32,9 @@ import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.comppot.mindsnack.R
 import com.comppot.mindsnack.model.Article
-import com.comppot.mindsnack.ui.components.ArticleCard
+import com.comppot.mindsnack.ui.components.ArticleItem
 import com.comppot.mindsnack.ui.components.CustomSearchBar
+import com.comppot.mindsnack.ui.components.EmptyListMessage
 import com.comppot.mindsnack.ui.components.FullScreenLoading
 import com.comppot.mindsnack.ui.components.SuggestBookDialog
 
@@ -58,7 +57,7 @@ fun SearchScreen(openArticle: (Long) -> Unit = {}, viewModel: SearchViewModel = 
         )
         when {
             state.isLoading -> FullScreenLoading()
-            state.articles.isEmpty() -> NoArticles()
+            state.articles.isEmpty() -> EmptyListMessage(stringResource(R.string.search_screen_no_articles))
             else -> ArticleList(state.articles, openArticle)
         }
     }
@@ -142,19 +141,8 @@ private fun ArticleList(articles: List<Article>, openArticle: (Long) -> Unit) {
         verticalArrangement = Arrangement.spacedBy(16.dp)
     ) {
         items(articles) {
-            ArticleCard(it, Modifier.fillMaxWidth(), openArticle)
+            ArticleItem(it, Modifier.fillMaxWidth(), openArticle)
         }
-    }
-}
-
-@Composable
-private fun NoArticles() {
-    Box(modifier = Modifier.fillMaxSize()) {
-        Text(
-            stringResource(id = R.string.search_screen_no_articles),
-            color = MaterialTheme.colorScheme.onSurfaceVariant,
-            modifier = Modifier.align(Alignment.Center)
-        )
     }
 }
